Update the design of src/data/source.py
For reasons still unclear, sometimes not all the images in a zip blob are extracted.  To ensure that modelling samples are extracted from an accessible set of images, the inventory file data is filtered to retain the records of accessible images only.

Additionally, the config.py import statement has been updated due to the re-location of src/config.py -> config.py

Modified: [src/data/source.py]